requirements.txt
boto3==1.25.4
botocore==1.28.4
jmespath==1.0.1
numpy==1.23.4
pandas==1.5.1
psycopg2==2.9.5
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.5
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12